extracted keyword call methods to KeywordCallObserver
